change merge behaviour for $packages
This is necessary in order to be able to replace a package
list. Otherwise changing a value could actually result in
having the old and new values being merged.

While here, prevent duplicate declaration errors in cases
where multiple components want to install the same package
that is used in bacula::client.